Performed preliminary nonnegative matrix factorization on the data
I have performed a preliminary analysis of the job descriptions with the nonnegative matrix factorization to identify common patterns in the job descriptions. For simplicity I am currently using only 1-grams and I have restricted myself to a subset of data. The pattern analysis is still ongoing, but it seems that the algorithm is capable to identify some broad classes of themes